pyproject.toml
[project]
name = "rag"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "docx2txt>=0.9",
    "fastapi>=0.135.2",
    "langchain>=1.2.13",
    "langchain-community>=0.4.1",
    "langchain-text-splitters>=1.1.1",
    "openai>=2.29.0",
    "pydantic>=2.12.5",
    "pyjwt>=2.12.1",
    "pypdf>=6.9.2",
    "python-dotenv>=1.2.2",
    "python-multipart>=0.0.22",
    "qdrant-client>=1.17.1",
    "tenacity>=9.1.4",
    "uvicorn>=0.42.0",
]
